api call will be able to get the game-servers location
